Revert "feat: add AbstractBroModule to support dependency cache inside the module"
This reverts commit 9bf76e43f26c22fea8a298dcf654de1810e5fbd0.

+++ sample/app.home/src/main/java/me/xx2bab/bro/sample/home/MainActivity.java
@@ -30,7 +30,7 @@ public class MainActivity extends AppCompatActivity {
     }
 
     public void getApi(View view) {
-        double pi = Bro.get().getApi(ISettingsApi.class).getPi();
+        int pi = Bro.get().getApi(ISettingsApi.class).getPi();
         Toast.makeText(this, pi + "", Toast.LENGTH_SHORT).show();
     }
